@@ -1,9 +1,12 @@
 package accounts.bank.managing.thesis.bachelor.rastvdmy.controller;
 
+import accounts.bank.managing.thesis.bachelor.rastvdmy.controller.mapper.UserMapper;
 import accounts.bank.managing.thesis.bachelor.rastvdmy.dto.request.UserRequest;
+import accounts.bank.managing.thesis.bachelor.rastvdmy.dto.response.UserResponse;
 import accounts.bank.managing.thesis.bachelor.rastvdmy.entity.User;
 import accounts.bank.managing.thesis.bachelor.rastvdmy.exception.ApplicationException;
 import accounts.bank.managing.thesis.bachelor.rastvdmy.service.UserService;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -18,234 +21,201 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.validation.Valid;
 import java.util.List;
+import java.util.UUID;
 
-/**
- * This class is responsible for handling user related requests.
- * It provides endpoints for getting all users, filtering users, getting a user by id,
- * creating a user, updating a user, uploading a user avatar,
- * updating a user's email, password, role, status, visibility, phone number,
- * and deleting a user.
- */
 @Slf4j
 @RestController
-@RequestMapping(path = "/profile")
+@RequestMapping(path = "/api/v1/users")
 public class UserController {
-
     private static final Logger LOG = LoggerFactory.getLogger(UserController.class);
     private final UserService userService;
+    private final UserMapper userMapper;
 
-    /**
-     * Constructor for the UserController.
-     *
-     * @param userService The service to handle user operations.
-     */
     @Autowired
-    public UserController(UserService userService) {
+    public UserController(UserService userService, UserMapper userMapper) {
         this.userService = userService;
+        this.userMapper = userMapper;
     }
 
-    /**
-     * This method is used to get all users.
-     *
-     * @return A list of all users.
-     */
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @ResponseStatus(HttpStatus.OK)
-    @GetMapping(path = "/")
-    @PreAuthorize("hasAnyRole('ROLE_MODERATOR', 'ROLE_ADMIN')")
-    public ResponseEntity<List<User>> getUsers() {
-        LOG.info("Getting all users ...");
-        return ResponseEntity.ok(userService.getUsers());
+    @GetMapping(produces = "application/json")
+    public ResponseEntity<List<UserResponse>> getUsers() {
+        logInfo("Getting all users ...");
+        List<User> users = userService.getUsers();
+        List<UserResponse> userResponses = users.stream().map(user -> userMapper.toResponse(
+                new UserRequest(
+                        user.getName(),
+                        user.getSurname(),
+                        user.getDateOfBirth(),
+                        user.getCountryOrigin(),
+                        user.getEmail(),
+                        user.getPassword(),
+                        user.getPhoneNumber()))
+        ).toList();
+        return ResponseEntity.ok(userResponses);
     }
 
-    /**
-     * This method is used to filter users.
-     *
-     * @param page The page number.
-     * @param size The size of the page.
-     * @param sort The sort order.
-     * @return A page of filtered users.
-     */
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @ResponseStatus(HttpStatus.OK)
-    @GetMapping(path = "/filter")
-    @PreAuthorize("hasAnyRole('ROLE_MODERATOR', 'ROLE_ADMIN')")
-    public ResponseEntity<Page<User>> filterUsers(
+    @GetMapping(path = "/filter", produces = "application/json")
+    public ResponseEntity<Page<UserResponse>> filterUsers(
             @RequestParam(value = "page", defaultValue = "0") int page,
             @RequestParam(value = "size", defaultValue = "10") int size,
             @RequestParam(value = "sort", defaultValue = "asc") String sort) {
-        LOG.info("Filtering users ...");
-        switch (sort.toLowerCase()) {
-            case "asc":
-                LOG.info("Sorting users by name in ascending order ...");
-                Pageable pageable = PageRequest.of(page, size, Sort.by("name").ascending());
-                return ResponseEntity.ok(userService.filterAndSortUsers(pageable));
-            case "desc":
-                LOG.info("Sorting users by name in descending order ...");
-                Pageable pageable1 = PageRequest.of(page, size, Sort.by("name").descending());
-                return ResponseEntity.ok(userService.filterAndSortUsers(pageable1));
-            default:
-                throw new ApplicationException(HttpStatus.BAD_REQUEST, "Invalid sort option. Use 'asc' or 'desc'.");
+        logInfo("Filtering users ...");
+        PageableState pageableState = parseSortOption(sort.trim().toLowerCase());
+        Pageable pageable = switch (pageableState) {
+            case ASC -> {
+                logInfo("Sorting users by name in ascending order ...");
+                yield PageRequest.of(page, size, Sort.by("name").ascending());
+            }
+            case DESC -> {
+                logInfo("Sorting users by name in descending order ...");
+                yield PageRequest.of(page, size, Sort.by("name").descending());
+            }
+        };
+        Page<User> users = userService.filterAndSortUsers(pageable);
+        Page<UserResponse> userResponses = users.map(user -> userMapper.toResponse(
+                new UserRequest(
+                        user.getName(),
+                        user.getSurname(),
+                        user.getDateOfBirth(),
+                        user.getCountryOrigin(),
+                        user.getEmail(),
+                        user.getPassword(),
+                        user.getPhoneNumber()))
+        );
+        return ResponseEntity.ok(userResponses);
+    }
+
+    private PageableState parseSortOption(String sortOption) {
+        try {
+            return PageableState.valueOf(sortOption.toUpperCase());
+        } catch (IllegalArgumentException e) {
+            throw new ApplicationException(HttpStatus.BAD_REQUEST,
+                    "Invalid sort option. Use 'sender' or 'receiver' and 'asc' or 'desc'.");
         }
     }
 
-    /**
-     * This method is used to get a user by id.
-     *
-     * @param userId The id of the user.
-     * @return The user with the given id.
-     */
+    @PreAuthorize("hasAnyRole('ROLE_DEFAULT', 'ROLE_ADMIN')")
     @ResponseStatus(HttpStatus.OK)
-    @GetMapping("/{id}")
-    @PreAuthorize("hasAnyRole('ROLE_MODERATOR', 'ROLE_ADMIN', 'ROLE_USER')")
-    public ResponseEntity<User> getUserById(@PathVariable(value = "id") Long userId) {
-        LOG.info("Getting user id: {} ...", userId);
-        return ResponseEntity.ok(userService.getUserById(userId));
+    @GetMapping(path = "/me", produces = "application/json")
+    public ResponseEntity<UserResponse> getMyUser(HttpServletRequest request) {
+        User user = userService.getUser(request);
+        UserResponse userResponse = userMapper.toResponse(
+                new UserRequest(
+                        user.getName(),
+                        user.getSurname(),
+                        user.getDateOfBirth(),
+                        user.getCountryOrigin(),
+                        user.getEmail(),
+                        user.getPassword(),
+                        user.getPhoneNumber()
+                )
+        );
+        return ResponseEntity.ok(userResponse);
     }
 
-    /**
-     * This method is used to create a user.
-     *
-     * @param user The request containing the user's name, surname, date of birth, country of origin, email, password, and phone number.
-     * @return The created user.
-     */
-    @ResponseStatus(HttpStatus.CREATED)
-    @PostMapping(path = "/register")
-    public ResponseEntity<User> createUser(@RequestBody UserRequest user) {
-        LOG.info("Creating user: {} ...", user.name());
-        return ResponseEntity.ok(userService.createUser(
-                user.name(),
-                user.surname(),
-                user.dateOfBirth(),
-                user.countryOfOrigin(),
-                user.email(),
-                user.password(),
-                user.phoneNumber()
-        ));
-    }
-
-    /**
-     * This method is used to update a user by id.
-     *
-     * @param userId The id of the user.
-     * @param user   The request containing the user's email, password, and phone number.
-     */
-    @ResponseStatus(HttpStatus.ACCEPTED)
-    @PatchMapping(path = "/{id}")
-    @PreAuthorize("hasAnyRole('ROLE_MODERATOR', 'ROLE_USER')")
-    public void updateUserById(@PathVariable(value = "id") Long userId, @RequestBody UserRequest user) {
-        LOG.info("Updating user id: {} ...", userId);
-        userService.updateUserById(userId,
-                user.email(),
-                user.password(),
-                user.phoneNumber());
-    }
-
-    /**
-     * This method is used to upload a user avatar.
-     *
-     * @param userId     The id of the user.
-     * @param userAvatar The user's avatar file.
-     */
-    @ResponseStatus(HttpStatus.ACCEPTED)
-    @PatchMapping(path = "/{id}/avatar")
-    @PreAuthorize("hasAnyRole('ROLE_MODERATOR', 'ROLE_USER')")
-    public void uploadUserAvatar(@PathVariable(value = "id") Long userId, @RequestBody MultipartFile userAvatar) {
-        LOG.info("Uploading user avatar id: {} ...", userId);
-        userService.uploadUserAvatar(userId, userAvatar);
-    }
-
-    /**
-     * This method is used to update a user's email by id.
-     *
-     * @param userId  The id of the user.
-     * @param request The request containing the user's email.
-     */
-    @ResponseStatus(HttpStatus.ACCEPTED)
-    @PatchMapping(path = "/{id}/email")
-    @PreAuthorize("hasAnyRole('ROLE_MODERATOR', 'ROLE_USER')")
-    public void updateUserEmailById(@PathVariable(value = "id") Long userId, @RequestBody UserRequest request) {
-        LOG.info("Updating user email id: {} ...", userId);
-        userService.updateUserEmailById(userId, request.email());
-    }
-
-    /**
-     * This method is used to update a user's password by id.
-     *
-     * @param userId  The id of the user.
-     * @param request The request containing the user's password.
-     */
-    @ResponseStatus(HttpStatus.ACCEPTED)
-    @PatchMapping(path = "/{id}/password")
-    @PreAuthorize("hasAnyRole('ROLE_MODERATOR', 'ROLE_USER')")
-    public void updateUserPasswordById(@PathVariable(value = "id") Long userId, @RequestBody UserRequest request) {
-        LOG.info("Updating user password id: {} ...", userId);
-        userService.updateUserPasswordById(userId, request.password());
-    }
-
-    /**
-     * This method is used to update a user's role by id.
-     *
-     * @param userId  The id of the user.
-     * @param request The request containing the user's role.
-     */
-    @ResponseStatus(HttpStatus.ACCEPTED)
-    @PatchMapping(path = "/{id}/role")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public void updateUserRoleById(@PathVariable(value = "id") Long userId, @RequestBody UserRequest request) {
-        LOG.info("Updating user role id: {} ...", userId);
-        userService.updateUserRoleById(userId, request.userRole());
+    @ResponseStatus(HttpStatus.OK)
+    @GetMapping(path = "/{id}", produces = "application/json")
+    public ResponseEntity<UserResponse> getUserInfo(@PathVariable("id") UUID id) {
+        logInfo("Getting user info ...");
+        User user = userService.getUserById(id);
+        UserResponse userResponse = userMapper.toResponse(
+                new UserRequest(
+                        user.getName(),
+                        user.getSurname(),
+                        user.getDateOfBirth(),
+                        user.getCountryOrigin(),
+                        user.getEmail(),
+                        user.getPassword(),
+                        user.getPhoneNumber()
+                )
+        );
+        return ResponseEntity.ok(userResponse);
     }
 
-    /**
-     * This method is used to update a user's status by id.
-     *
-     * @param userId The id of the user.
-     */
-    @ResponseStatus(HttpStatus.ACCEPTED)
-    @PatchMapping(path = "/{id}/status")
-    @PreAuthorize("hasAnyRole('ROLE_MODERATOR', 'ROLE_ADMIN')")
-    public void updateUserStatusById(@PathVariable(value = "id") Long userId) {
-        LOG.info("Updating user state id: {} ...", userId);
-        userService.updateUserStatusById(userId);
+    @PreAuthorize("hasRole('ROLE_DEFAULT')")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    @PatchMapping(path = "/me", consumes = "application/json", produces = "application/json")
+    public ResponseEntity<UserResponse> updateUser(HttpServletRequest request,
+                                                   @Valid @RequestBody UserRequest userRequest) throws Exception {
+        logInfo("Updating user: {} ...", userRequest.name());
+        User user = userService.updateUser(request, userRequest);
+        UserResponse userResponse = userMapper.toResponse(
+                new UserRequest(
+                        user.getName(),
+                        user.getSurname(),
+                        user.getDateOfBirth(),
+                        user.getCountryOrigin(),
+                        user.getEmail(),
+                        user.getPassword(),
+                        user.getPhoneNumber()
+                )
+        );
+        return ResponseEntity.ok(userResponse);
     }
 
-    /**
-     * This method is used to update a user's visibility by id.
-     *
-     * @param userId The id of the user.
-     */
-    @ResponseStatus(HttpStatus.ACCEPTED)
-    @PatchMapping(path = "/{id}/visibility")
-    @PreAuthorize("hasAnyRole('ROLE_MODERATOR', 'ROLE_ADMIN', 'ROLE_USER')")
-    public void updateUserVisibilityById(@PathVariable(value = "id") Long userId) {
-        LOG.info("Updating user visibility id: {} ...", userId);
-        userService.updateUserVisibilityById(userId);
+    @PreAuthorize("hasRole('ROLE_DEFAULT')")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    @PatchMapping(path = "/me/avatar", consumes = "application/json", produces = "application/json")
+    public ResponseEntity<Void> manageUserAvatar(HttpServletRequest request,
+                                                 @RequestParam("action") String action,
+                                                 @RequestBody(required = false) MultipartFile userAvatar) {
+        if ("upload".equalsIgnoreCase(action)) {
+            logInfo("Uploading user avatar...");
+            userService.uploadUserAvatar(request, userAvatar);
+        } else if ("remove".equalsIgnoreCase(action)) {
+            logInfo("Deleting user avatar...");
+            userService.removeUserAvatar(request);
+        } else {
+            throw new ApplicationException(HttpStatus.BAD_REQUEST, "Invalid action specified.");
+        }
+        return ResponseEntity.noContent().build();
     }
 
-    /**
-     * This method is used to update a user's phone number by id.
-     *
-     * @param userId  The id of the user.
-     * @param request The request containing the user's phone number.
-     */
-    @ResponseStatus(HttpStatus.ACCEPTED)
-    @PatchMapping(path = "/{id}/phone-number")
-    @PreAuthorize("hasAnyRole('ROLE_MODERATOR', 'ROLE_USER')")
-    public void updateUserPhoneNumberById(@PathVariable(value = "id") Long userId, @RequestBody UserRequest request) {
-        LOG.info("Updating user phone number id: {} ...", userId);
-        userService.updateUserPhoneNumberById(userId, request.phoneNumber());
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    @PatchMapping(path = "/role/{id}", consumes = "application/json", produces = "application/json")
+    public ResponseEntity<Void> updateUserRole(@PathVariable(value = "id") UUID id) {
+        logInfo("Granting user new role ...");
+        userService.updateUserRole(id);
+        return ResponseEntity.noContent().build();
     }
 
-    /**
-     * This method is used to delete a user by id.
-     *
-     * @param userId The id of the user.
-     */
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    @PatchMapping(path = "/status/{id}", consumes = "application/json", produces = "application/json")
+    public ResponseEntity<Void> updateUserStatus(@PathVariable("id") UUID id,
+                                                 @RequestParam(value = "status") String status) {
+        logInfo("Updating user status to: {} ...", status);
+        userService.updateUserStatus(id, status);
+        return ResponseEntity.noContent().build();
+    }
+
+    @PreAuthorize("hasRole('ROLE_DEFAULT')")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    @DeleteMapping(path = "/me")
+    public ResponseEntity<Void> deleteMyProfile(HttpServletRequest request) {
+        logInfo("Deleting user ...");
+        userService.deleteUser(request);
+        return ResponseEntity.noContent().build();
+    }
+
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping(path = "/{id}")
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public void deleteUserById(@PathVariable(value = "id") Long userId) {
-        LOG.info("Deleting user by id: {} ...", userId);
-        userService.deleteUserById(userId);
+    public ResponseEntity<Void> deleteUserProfile(@PathVariable("id") UUID id) {
+        logInfo("Deleting user ...");
+        userService.deleteUserById(id);
+        return ResponseEntity.noContent().build();
+    }
+
+    private void logInfo(String message, Object... args) {
+        LOG.info(message, args);
     }
 }
