@@ -1,7 +1,11 @@
 package accounts.bank.managing.thesis.bachelor.rastvdmy.controller;
 
+import accounts.bank.managing.thesis.bachelor.rastvdmy.controller.mapper.CurrencyDataMapper;
+import accounts.bank.managing.thesis.bachelor.rastvdmy.dto.request.CurrencyDataRequest;
+import accounts.bank.managing.thesis.bachelor.rastvdmy.dto.response.CurrencyDataResponse;
 import accounts.bank.managing.thesis.bachelor.rastvdmy.entity.CurrencyData;
 import accounts.bank.managing.thesis.bachelor.rastvdmy.service.CurrencyDataService;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -12,55 +16,57 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
-/**
- * This class is responsible for handling currency data related requests.
- * It provides endpoints for updating and fetching all currencies and finding a currency by its type.
- */
 @Slf4j
 @RestController
-@RequestMapping("/api/currency-data")
+@PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_DEFAULT')")
+@RequestMapping("/api/v1/currency-data")
 public class CurrencyDataController {
     private static final Logger LOG = LoggerFactory.getLogger(CurrencyDataController.class);
 
     private final CurrencyDataService currencyDataService;
+    private final CurrencyDataMapper currencyDataMapper;
 
-    /**
-     * Constructor for the CurrencyDataController.
-     *
-     * @param currencyDataService The service to handle currency data operations.
-     */
     @Autowired
-    public CurrencyDataController(CurrencyDataService currencyDataService) {
+    public CurrencyDataController(CurrencyDataService currencyDataService, CurrencyDataMapper currencyDataMapper) {
         this.currencyDataService = currencyDataService;
+        this.currencyDataMapper = currencyDataMapper;
     }
 
-    /**
-     * This method is used to update and fetch all currencies.
-     *
-     * @return A list of all currency data.
-     */
     @ResponseStatus(HttpStatus.OK)
-    @GetMapping(path = "/")
-    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_USER', 'ROLE_MODERATOR')")
-    public ResponseEntity<List<CurrencyData>> updateAndFetchAllCurrencies() {
-        LOG.info("Updating currency data ...");
+    @GetMapping(produces = "application/json")
+    public ResponseEntity<List<CurrencyDataResponse>> updateAndFetchAllCurrencies(HttpServletRequest request) {
+        logInfo("Updating currency data ...");
         currencyDataService.findAllExchangeRates();
-        return ResponseEntity.ok(currencyDataService.findAllCurrencies());
+        List<CurrencyData> currencies = currencyDataService.findAllCurrencies(request);
+        List<CurrencyDataResponse> currencyDataResponses = currencies.stream()
+                .map(currencyData -> currencyDataMapper.toResponse(new CurrencyDataRequest(
+                                currencyData.getId(),
+                                currencyData.getCurrency(),
+                                currencyData.getRate()
+                        ))
+                ).collect(Collectors.toList());
+        return ResponseEntity.ok(currencyDataResponses);
     }
 
-    /**
-     * This method is used to find a currency by its type.
-     *
-     * @param currencyType The type of the currency.
-     * @return The currency data of the given type.
-     */
     @ResponseStatus(HttpStatus.OK)
-    @GetMapping(path = "/{currency}")
-    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_USER', 'ROLE_MODERATOR')")
-    public ResponseEntity<CurrencyData> findByCurrency(@PathVariable(value = "currency") String currencyType) {
-        LOG.info("Getting currency {} ...", currencyType);
+    @GetMapping(path = "/{currency}", produces = "application/json")
+    public ResponseEntity<CurrencyDataResponse> findByCurrency(
+            HttpServletRequest request,
+            @PathVariable(value = "currency") String currencyType) {
+        logInfo("Getting currency {} ...", currencyType);
         currencyDataService.findAllExchangeRates();
-        return ResponseEntity.ok(currencyDataService.findByCurrency(currencyType));
+        CurrencyData currencyData = currencyDataService.findByCurrency(request, currencyType);
+        CurrencyDataResponse currencyDataResponse = currencyDataMapper.toResponse(new CurrencyDataRequest(
+                currencyData.getId(),
+                currencyData.getCurrency(),
+                currencyData.getRate()
+        ));
+        return ResponseEntity.ok(currencyDataResponse);
+    }
+
+    private void logInfo(String message, Object... args) {
+        LOG.info(message, args);
     }
 }
