@@ -1,8 +1,16 @@
 package accounts.bank.managing.thesis.bachelor.rastvdmy.service.component;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.hazelcast.spi.exception.RestClientException;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
+import org.springframework.web.client.RestTemplate;
 
+import java.io.IOException;
+import java.time.LocalDate;
 import java.util.Random;
+import java.util.concurrent.ThreadLocalRandom;
 
 /**
  * This class is responsible for generating various types of random strings.
@@ -10,33 +18,20 @@ import java.util.Random;
  */
 @Component
 public class Generator {
+    private final RestTemplate restTemplate;
 
-    /**
-     * Generates a reference number consisting of 8 random uppercase letters followed by 1 to 3 random digits.
-     *
-     * @return The generated reference number.
-     */
-    public String generateReferenceNumber() {
-        String referenceNumber;
-        StringBuilder sb = new StringBuilder();
+    public Generator(RestTemplate restTemplate) {
+        this.restTemplate = restTemplate;
+    }
+
+    public String generateBankNumber() {
         Random random = new Random();
-
-        // Generate 8 random uppercase letters
-        for (int i = 0; i < 8; i++) {
-            char randomChar = (char) (random.nextInt(26) + 'A');
-            sb.append(randomChar);
+        StringBuilder bankNumber = new StringBuilder();
+        // Generate 4 random digits for the bank number
+        for (int i = 0; i < 4; i++) {
+            bankNumber.append(random.nextInt(10));
         }
-
-        // Determine the number of digits to append
-        int numDigits = random.nextInt(3) + 1;
-
-        // Generate the determined number of random digits
-        for (int i = 0; i < numDigits; i++) {
-            int randomDigit = random.nextInt(10);
-            sb.append(randomDigit);
-        }
-        referenceNumber = sb.toString();
-        return referenceNumber;
+        return bankNumber.toString();
     }
 
     /**
@@ -63,6 +58,26 @@ public class Generator {
             iban.append(random.nextInt(10));
         }
         return iban.toString();
+    }
+
+    public String generateCvv() {
+        Random random = new Random();
+        StringBuilder cvv = new StringBuilder();
+        // Generate 3 random digits for the CVV code
+        for (int i = 0; i < 3; i++) {
+            cvv.append(random.nextInt(10));
+        }
+        return cvv.toString();
+    }
+
+    public String generatePin() {
+        Random random = new Random();
+        StringBuilder pin = new StringBuilder();
+        // Generate 4 random digits for the PIN code
+        for (int i = 0; i < 4; i++) {
+            pin.append(random.nextInt(10));
+        }
+        return pin.toString();
     }
 
     /**
@@ -94,7 +109,66 @@ public class Generator {
         for (int i = 0; i < 10; i++) {
             accountNumber.append(random.nextInt(10));
         }
-        accountNumber.append("/0800");
         return accountNumber.toString();
+    }
+
+    public String generateCardNumber() {
+        Random random = new Random();
+        StringBuilder cardNumber = new StringBuilder();
+        // Generate 16 random digits for the card number
+        for (int i = 0; i < 16; i++) {
+            cardNumber.append(random.nextInt(10));
+        }
+        return cardNumber.toString();
+    }
+
+    protected boolean countryExists(String countryName) {
+        final String url = "https://restcountries.com/v3.1/all?fields=name";
+        try {
+            ResponseEntity<String> response = restTemplate.getForEntity(url, String.class);
+            if (response.getStatusCode().is2xxSuccessful() && response.getBody() != null) {
+                ObjectMapper mapper = new ObjectMapper();
+                JsonNode rootNode = mapper.readTree(response.getBody());
+                if (rootNode.isArray()) {
+                    for (JsonNode node : rootNode) {
+                        JsonNode nameNode = node.get("name");
+                        if (nameNode != null && nameNode.get("common") != null) {
+                            String commonName = nameNode.get("common").asText();
+                            if (commonName.equalsIgnoreCase(countryName)) {
+                                return false;
+                            }
+                        }
+                    }
+                }
+            }
+        } catch (RestClientException | IOException e) {
+            e.getCause();
+        }
+        return true;
+    }
+
+    public static LocalDate getRandomStartDate() {
+        int currentYear = LocalDate.now().getYear();
+
+        int startYear = currentYear - 3;
+
+        long startEpochDay = LocalDate.of(startYear, 1, 1).toEpochDay();
+        long endEpochDay = LocalDate.of(currentYear, 12, 31).toEpochDay();
+
+        long randomDay = ThreadLocalRandom.current().nextLong(startEpochDay, endEpochDay + 1);
+        return LocalDate.ofEpochDay(randomDay);
+    }
+
+    public static LocalDate getRandomExpirationDate(LocalDate startDate) {
+        LocalDate minExpirationDate = startDate.plusYears(3); // 3 years from the start date
+        LocalDate maxExpirationDate = startDate.plusYears(5); // 5 years from the start date
+
+        // Convert these dates to epoch days
+        long startEpochDay = minExpirationDate.toEpochDay();
+        long endEpochDay = maxExpirationDate.toEpochDay();
+
+        // Generate a random expiration date within this range
+        long randomDay = ThreadLocalRandom.current().nextLong(startEpochDay, endEpochDay + 1);
+        return LocalDate.ofEpochDay(randomDay);
     }
 }
